move json-rpc to client/
